feat: bundle merchant_classifier.tflite + ML layer bug fixes
- Add merchant_classifier.tflite (256→128→64→13 softmax, 49KB) to
  assets/models/ so MLCategoryClassifier activates without a model
  trained externally; update .gitignore to allow assets/models/
- BehavioralCluster: guard empty-expenses case in extractFeatures()
  (prevented NaN feature vectors when user has income-only data);
  make classify() suspend and add Mutex around Interpreter.run()
- MLCategoryClassifier: add Mutex around Interpreter.run() — TFLite
  Interpreter is not thread-safe under concurrent coroutine calls
- SpendingPredictor: same Mutex fix; make predict() suspend to match
- TextFeatureExtractor: remove unnecessary Double→Float→Double roundtrip
  in L2 norm computation (precision loss in sqrt input)

+++ app/src/main/kotlin/com/financeos/hub/core/ml/TextFeatureExtractor.kt
@@ -34,7 +34,7 @@ class TextFeatureExtractor @Inject constructor() {
         }
 
         // L2 normalise
-        val norm = kotlin.math.sqrt(vector.sumOf { (it * it).toDouble() }.toFloat().toDouble()).toFloat()
+        val norm = kotlin.math.sqrt(vector.sumOf { (it * it).toDouble() }).toFloat()
         if (norm > 0f) {
             for (i in vector.indices) vector[i] /= norm
         }
